# src/main/resources/openapi/api/api.yaml
openapi: 3.0.1
info:
  title: Cocktail API
  version: LATEST
servers:
  - url: http://localhost:8080/
paths:
...
  /cocktails:
    get:
      tags:
        - cocktail
      operationId: getCocktails
      parameters:
        - in: query
          name: start
          required: false
          schema:
            type: integer
        - in: query
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CocktailsLightResponse'
    post:
      tags:
        - cocktail
      operationId: saveCocktail
      parameters:
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCocktailRequestDTO'
      responses:
        200:
          description: OK
  /v2/cocktails:
    get:
      tags:
        - cocktail
      operationId: getCocktailsByName
      parameters:
        - in: query
          name: start
          required: false
          schema:
            type: integer
        - in: query
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCocktailLightResponse'
  /cocktails/ingredients/{id}:
    get:
      tags:
        - cocktail
      operationId: getCocktailsByIngredient
      parameters:
        - in: query
          name: start
          required: false
          schema:
            type: integer
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CocktailsLightResponse'
  /cocktails/ingredients:
    get:
      tags:
        - cocktail
      operationId: getCocktailsByIngredients
      parameters:
        - in: query
          name: start
          required: false
          schema:
            type: integer
        - in: query
          name: ingredientIds
          required: true
          schema:
            type: array
            items:
              type: integer
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CocktailsLightResponse'
  /v2/cocktails/ingredients/search:
    get:
      tags:
        - cocktail
      operationId: getAllCocktailsByIngredients
      parameters:
        - in: query
          name: start
          required: false
          schema:
            type: integer
        - in: query
          name: ingredientIds
          required: true
          schema:
            type: array
            items:
              type: integer
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCocktailLightResponse'
  /cocktails/{id}:
    get:
      tags:
        - cocktail
      operationId: getCocktail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CocktailDTO'
  /cocktails-by-id:
    get:
      tags:
        - cocktail
      operationId: getCocktailsByIds
      parameters:
        - in: query
          name: ids
          required: true
          schema:
            type: array
            items:
              type: integer
        - in: query
          name: start
          required: false
          schema:
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CocktailsLightResponse'
  /cocktails/names/{name}:
    get:
      tags:
        - cocktail
      operationId: getCocktailNames
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CocktailNamesResponse'
  /ingredients/names/{name}:
    get:
      tags:
        - ingredient
      operationId: getIngredientNames
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientNamesResponse'
  /ingredients:
    get:
      tags:
        - ingredient
      operationId: getIngredients
      parameters:
        - in: query
          name: start
          required: false
          schema:
            type: integer
        - in: query
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientsLightResponse'
  /ingredients-by-id:
    get:
      tags:
        - ingredient
      operationId: getIngredientsByIds
      parameters:
        - in: query
          name: ids
          required: true
          schema:
            type: array
            items:
              type: integer
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientsLightResponse'
  /ingredients/{id}:
    get:
      tags:
        - ingredient
      operationId: getIngredient
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientDTO'
  /ingredients/types:
    get:
      tags:
        - ingredient
      operationId: getIngredientTypes
      parameters:
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientsLightResponse'
  /tools/types:
    get:
      tags:
        - tool
      operationId: getTools
      parameters:
        - in: header
          name: requestId
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolsLightResponse'
components:
  schemas:
...
    CocktailNamesResponse:
      type: array
      items:
        $ref: '#/components/schemas/CocktailNameDTO'
    CocktailNameDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    IngredientNamesResponse:
      type: array
      items:
        $ref: '#/components/schemas/IngredientNameDTO'
    IngredientNameDTO:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          $ref: '#/components/schemas/IngredientTypeDTO'
    CocktailsLightResponse:
      type: array
      items:
        $ref: '#/components/schemas/CocktailLightDTO'
    PagedCocktailLightResponse:
      type: object
      properties:
        hasNext:
          type: boolean
        nextStart:
          type: integer
        cocktails:
          type: array
          items:
            $ref: '#/components/schemas/CocktailLightDTO'
    CocktailLightDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        preview:
          type: string
          format: byte
        missingIngredientsCount:
          type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/CocktailIngredientDTO'
        isCustom:
          type: boolean
    CocktailIngredientDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        preview:
          type: string
          format: byte
        type:
          $ref: '#/components/schemas/IngredientTypeDTO'
        amount:
          type: integer
        unit:
          type: string
    CocktailDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
          format: byte
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/CocktailIngredientDTO'
        tools:
          type: array
          items:
            $ref: '#/components/schemas/CocktailToolDTO'
        instructions:
          type: array
          items:
            type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        isCustom:
          type: boolean
    SaveCocktailRequestDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
          format: byte
        preview:
          type: string
          format: byte
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/SaveIngredientRequestDTO'
        tools:
          type: array
          items:
            type: integer
        instructions:
          type: array
          items:
            type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    SaveIngredientRequestDTO:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: integer
        unit:
          type: string
    CocktailToolDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        preview:
          type: string
          format: byte
    IngredientDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
          format: byte
        type:
          $ref: '#/components/schemas/IngredientTypeDTO'
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    IngredientsLightResponse:
      type: array
      items:
        $ref: '#/components/schemas/IngredientLightDTO'
    ToolsLightResponse:
      type: array
      items:
        $ref: '#/components/schemas/ToolLightDTO'
    IngredientLightDTO:
      type: object
      properties:
        id:
          type: integer
        preview:
          type: string
          format: byte
        name:
          type: string
        type:
          $ref: '#/components/schemas/IngredientTypeDTO'
    ToolLightDTO:
      type: object
      properties:
        id:
          type: integer
        preview:
          type: string
          format: byte
        name:
          type: string
...
    IngredientTypeDTO:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - GIN
            - VODKA
            - TEQUILA
            - RUM
            - WHISKEY
            - BOURBON
            - ABSINTHE
            - COGNAC
            - PORT_WINE
            - SHERRY
            - APERITIF
            - LIQUOR
            - VERMOUTH
            - BITTER
            - WINE
            - BEER
            - SYRUP
            - NON_ALCOHOLIC
            - HONEY
            - DESSERT
            - ADDITIVE
            - GREENS
            - MISC
            - JAM
            - JUICE
            - SPICE
            - DRIED_FRUIT
            - DECORATION
            - FRUIT
            - VEGETABLE
            - TEA
            - NUT
            - BERRY
            - ICE